refactor: extract FlopWeightsTreeView into its own module
The tree-rendering of flop weights and the loading/aggregation of the built-in data are two separate concerns that shared _builtin_data.py; split the view class into _flop_weights_tree_view.py so the data module holds only data logic.
